requirements.txt
asgiref==3.3.1
Django==3.1.8
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
mssql-django==1.0b1
PyJWT==2.0.1
pyodbc==4.0.30
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
